updating api to accept date filters and filtering responses from neo4j

+++ src/main/java/bitcoin/spring/data/neo4j/domain/relationships/InputRelation.java
@@ -50,7 +50,7 @@ public class InputRelation {
         return inputValue * eurExchangeRate;
     }
 
-    public double getTimestamp() {
+    public long getTimestamp() {
         return transaction.getMinedInBlock().getTimestamp();
     }
 }
